pyproject.toml

[tool.poetry]
name = "ai-travel-agent"
version = "0.1.1"
description = "AI-powered travel planning agent with flight and hotel search capabilities"
authors = ["Nir Bar"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
langchain = "^0.2.0"
langchain-google-genai = "^1.0.0"
langgraph = "^0.2.0"
grandalf = "^0.8"
mailgun = "^0.1.1"
serpapi = "^0.1.5"
beautifulsoup4 = "^4.12.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
